# openapi.yaml
openapi: 3.0.3
info:
  title: University Services API
  version: 1.0.0
  description: 'The central API platform for university operations.

    Covers Authentication, Students, Programs & Courses, Faculty, Library, AI
    Assistant, API Key Management, and UAE Government Reporting domains.


    ## Authentication

    Most endpoints require a Bearer JWT token obtained via `POST /auth/login`.

    Government reporting endpoints (`/gov/*`) require an `X-API-Key` header instead.


    Get a JWT token: `POST /auth/login` with your email and password.

    Use JWT: Add header `Authorization: Bearer <token>` to all requests.

    Use API Key: Add header `X-API-Key: uak_<your_key>` — service role only.


    ## Test Credentials (local / seeded environment)

    | Role | Email | Password |

    |------|-------|----------|

    | Admin | admin@university.ac.ae | Admin@1234 |

    | Student | student1@university.ac.ae | Student@1234 |

    | Faculty | faculty1@university.ac.ae | Faculty@1234 |

    | Librarian | librarian@university.ac.ae | Librarian@1234 |


    ## API-Led Connectivity Architecture

    Follows MuleSoft API-led connectivity:

    - **System Layer**: Raw data access (student records, course catalog, book inventory)

    - **Process Layer**: Business logic (enrollment validation, borrowing rules, capacity
    checks)

    - **Experience Layer**: Consumer-tailored responses (student portal, faculty dashboard)

...


    ## AI Library Assistant

...
    OpenAI.

...

    '
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via OAuth2 authorization flow
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100
      description: Number of results per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: 'Field to sort by. Prefix with - for descending. Example: sort=-created_at'
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 25
        total_items:
          type: integer
          example: 142
        total_pages:
          type: integer
          example: 6
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: The request body contains invalid fields
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  issue:
                    type: string
              example:
              - field: email
                issue: Must be a valid university email address
...
    BookBase:
      type: object
      required:
      - title
      - author
      - isbn
      - genre
      - total_copies
      properties:
        title:
          type: string
          example: Introduction to Algorithms
        author:
          type: string
          example: Thomas Cormen
        isbn:
          type: string
          example: '9780262033848'
        genre:
          type: string
          example: computer_science
        total_copies:
          type: integer
          example: 5
        publisher:
          type: string
          example: MIT Press
        published_year:
          type: integer
          example: 2009
    Book:
      allOf:
      - $ref: '#/components/schemas/BookBase'
      - type: object
        properties:
          id:
            type: string
            example: book_301
            readOnly: true
          available_copies:
            type: integer
            description: Server-computed. Read-only — calculated from total_copies
              minus active borrowings.
            readOnly: true
            example: 3
          created_at:
            type: string
            format: date-time
            readOnly: true
          updated_at:
            type: string
            format: date-time
            readOnly: true
    BookListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Book'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    BorrowingRequest:
      type: object
      required:
      - book_id
      - student_id
      - due_date
      properties:
        book_id:
          type: string
          example: book_301
        student_id:
          type: string
          example: stu_10042
        due_date:
          type: string
          format: date
          example: '2026-03-25'
    BorrowingUpdate:
      type: object
      properties:
        status:
          type: string
          enum:
          - returned
          - lost
          description: 'Updating status to ''returned'' automatically sets returned_at
            to current timestamp.

            The client never sends returned_at directly — it is server-managed.

            '
    Borrowing:
      type: object
      properties:
        id:
          type: string
          example: brw_60001
          readOnly: true
        book_id:
          type: string
          example: book_301
        student_id:
          type: string
          example: stu_10042
        borrowed_at:
          type: string
          format: date-time
          readOnly: true
        due_date:
          type: string
          format: date
          example: '2026-03-25'
        returned_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          description: Set automatically by server when status changes to 'returned'
        status:
          type: string
          enum:
          - active
          - returned
          - overdue
          - lost
          example: active
    BorrowingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Borrowing'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict with current resource state (e.g. duplicate email, already
        enrolled)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Semantically invalid — passes validation but violates a business
        rule (e.g. course at capacity)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
- BearerAuth: []
paths:
...
  /books:
    get:
      tags:
      - Library - Books
      summary: List all books
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/SortParam'
      - name: genre
        in: query
        schema:
          type: string
        example: computer_science
      - name: available
        in: query
        schema:
          type: boolean
        description: Filter to only books with available copies
      - name: author
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
        description: Full-text search across title, author, and description
      responses:
        '200':
          description: Paginated list of books
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: listBooks
    post:
      tags:
      - Library - Books
      summary: Add a new book to the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookBase'
      responses:
        '201':
          description: Book added to catalog
          headers:
            Location:
              schema:
                type: string
                example: /books/book_301
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
      operationId: createBook
  /books/{book_id}:
    parameters:
    - name: book_id
      in: path
      required: true
      schema:
        type: string
      example: book_301
    get:
      tags:
      - Library - Books
      summary: Get a specific book
      responses:
        '200':
          description: Book record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getBook
    put:
      tags:
      - Library - Books
      summary: Full update of a book record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookBase'
      responses:
        '200':
          description: Book updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: replaceBook
    patch:
      tags:
      - Library - Books
      summary: Partial update (e.g. update total_copies when new copies arrive)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                total_copies:
                  type: integer
                genre:
                  type: string
      responses:
        '200':
          description: Book partially updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: updateBook
    delete:
      tags:
      - Library - Books
      summary: Remove a book from the catalog
      responses:
        '204':
          description: Book deleted
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: deleteBook
  /books/{book_id}/borrowings:
    get:
      tags:
      - Library - Books
      summary: Get borrowing history for a specific book
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      - name: status
        in: query
        schema:
          type: string
          enum:
          - active
          - returned
          - overdue
          - lost
      responses:
        '200':
          description: Paginated borrowing history for this book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowingListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getBookBorrowings
  /borrowings:
    get:
      tags:
      - Library - Borrowings
      summary: List all borrowing records
      description: "Top-level resource — not nested under books or students because\
        \ \nborrowings are relationships between both, queryable from either direction.\n"
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      - name: student_id
        in: query
        schema:
          type: string
        description: Filter borrowings by student
        example: stu_10042
      - name: book_id
        in: query
        schema:
          type: string
        description: Filter borrowings by book
      - name: status
        in: query
        schema:
          type: string
          enum:
          - active
          - returned
          - overdue
          - lost
        description: Filter by borrowing status. Use status=overdue for library overdue
          reports.
      responses:
        '200':
          description: Paginated list of borrowing records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowingListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: listBorrowings
    post:
      tags:
      - Library - Borrowings
      summary: Borrow a book
      description: 'Business rules enforced:

        - Book must exist and have available copies

        - Student must exist and have active account

        - Student must not already have this book borrowed

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowingRequest'
      responses:
        '201':
          description: Book borrowed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Borrowing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
      operationId: createBorrowing
  /borrowings/{borrowing_id}:
    parameters:
    - name: borrowing_id
      in: path
      required: true
      schema:
        type: string
      example: brw_60001
    get:
      tags:
      - Library - Borrowings
      summary: Get a specific borrowing record
      responses:
        '200':
          description: Borrowing record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Borrowing'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getBorrowing
    patch:
      tags:
      - Library - Borrowings
      summary: Update a borrowing — primarily used to return a book
      description: 'Setting status to ''returned'' automatically triggers:

        - returned_at set to current server timestamp

        - available_copies on the book incremented by 1

        The client never manages these side effects — they are server-handled.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowingUpdate'
      responses:
        '200':
          description: Borrowing updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Borrowing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: updateBorrowing
...